web_app/requirements.txt
accelerate==0.31.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
async-timeout==5.0.1
attrs==25.3.0
aws-xray-sdk
blinker==1.9.0
boto3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crcmod==1.7
cryptography==44.0.2
datasets==2.12.0
diffusers==0.29.0
dill==0.3.6
einops==0.8.1
filelock==3.18.0
Flask==3.1.0
frozenlist==1.6.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
lightning-utilities==0.14.3
MarkupSafe==3.0.2
modelscope==1.9.5
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.14
networkx==3.4.2
numpy==1.26.4
open_clip_torch==2.32.0
opencv-python==4.11.0.86
oss2==2.19.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-dotenv
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sortedcontainers==2.4.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.19.1
tomli==2.2.1
torch>=2.1.0
torchaudio
torchdata
torchmetrics
torchtext
torchvision
tqdm==4.67.1
transformers==4.41.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.0
zipp==3.21.0
